node_type can be 16-bits

+++ packages/db/clibs/include/selva.h
@@ -37,7 +37,7 @@ struct SelvaObject;
 
 typedef int8_t field_t;
 typedef uint32_t node_id_t;
-typedef uint32_t node_type_t;
+typedef uint16_t node_type_t;
 
 RB_HEAD(SelvaNodeIndex, SelvaNode);
 RB_HEAD(SelvaTypeIndex, SelvaTypeEntry);
